requirement.txt
flask
opencv-python
google-generativeai
mysql-connector-python
python-dotenv